Abstraction layer for MarketOperatorDataStorageService.cs to let tests run.
